Move styling to Paragraph component, to avoid...
...bloating renderer.ts
